Wrong blur radius for filter: drop-shadow() https://bugs.webkit.org/show_bug.cgi?id=163667
Reviewed by Darin Adler.

Source/WebCore:

The "stdDev" parameter to <feDropShadow> and <feGaussianBlur> represents the standard
deviation (σ) of the Gaussian blur function. CSS filters drop-shadow() and blur() follow
this with their radius parameter.

However, CSS box-shadow and text-shadow use 2σ to describe the blur radius, since this
is conveniently close to the number of pixels the effect of the blur extends out.

feDropShadow, which is used by non-accelerated filter: drop-shadow(), was using
the wrong blur radius because it passed its stdDev directly to ShadowBlur. ShadowBlur
was written for CSS box-shadow, so expected its input "blur radius" to be twice the
stdDev.

Tests: css3/filters/drop-shadow-blur-radius.html
       svg/filters/feDropShadow-blur-radius.html

* platform/graphics/filters/FEDropShadow.cpp:
(WebCore::FEDropShadow::platformApplySoftware):

LayoutTests:

Mismatch tests to detect that the blurry area extends out from under a masking
element on top.

* css3/filters/drop-shadow-blur-radius-expected-mismatch.html: Added.
* css3/filters/drop-shadow-blur-radius.html: Added.
* svg/filters/feDropShadow-blur-radius-expected-mismatch.html: Added.
* svg/filters/feDropShadow-blur-radius.html: Added.


Identifier: 181843@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@208058 268f45cc-cd09-0410-ab3c-d52691b4dbfc