AX: iOS Voiceover does not announce previously selected value from input type="date" form field https://bugs.webkit.org/show_bug.cgi?id=164176
Reviewed by Chris Fleizach.

Source/WebCore:

Input type date is a popup button on iOS and its value was not exposed in stringValue() since
it's not considered a text control. Instead, the value was exposed in AXTitle. Fixed this by adding
the case in stringValue() and removing the AXTitle exposure.

Changes are covered in modified test.

* accessibility/AccessibilityObject.h:
* accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::stringValue):
* accessibility/ios/AccessibilityObjectIOS.mm:
(WebCore::AccessibilityObject::isInputTypePopupButton):
* accessibility/ios/WebAccessibilityObjectWrapperIOS.mm:
(-[WebAccessibilityObjectWrapper accessibilityLabel]):

LayoutTests:

* accessibility/ios-simulator/input-type-time-expected.txt:
* accessibility/ios-simulator/input-type-time.html:


Identifier: 181940@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@208159 268f45cc-cd09-0410-ab3c-d52691b4dbfc