[css-grid] mimax(auto, <flex>) should be serialized as <flex> https://bugs.webkit.org/show_bug.cgi?id=164316
Reviewed by Sergio Villar Senin.

Source/WebCore:

* css/CSSComputedStyleDeclaration.cpp:
(WebCore::specifiedValueForGridTrackSize): Add a simple if to serialize
properly this case.
* rendering/style/GridLength.h:
(WebCore::GridLength::isAuto): Add new function to check if GridLength
is or not "auto".

LayoutTests:

Add new test case to verify it. We can only check it
using grid-auto-columns|rows, because grid-template-columns|rows
is serialized to the used breadth.

* fast/css-grid-layout/grid-auto-columns-rows-get-set-expected.txt:
* fast/css-grid-layout/grid-auto-columns-rows-get-set.html:


Identifier: 182037@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@208277 268f45cc-cd09-0410-ab3c-d52691b4dbfc