CSS.supports("font-variation-settings", "'wght' 500") erroneously returns false https://bugs.webkit.org/show_bug.cgi?id=164244
Reviewed by Michael Catanzaro.

Source/WebCore:

Because we weren't passing a Document into CSSParserContext's constructor,
there was no way for our parser to know whether the runtime switch was on
or not. Instead, we can use the CallWith attribute in the IDL file to pass
in a Document.

Test: fast/text/variations/css-supports-runtime-switch.html

* css/DOMCSSNamespace.cpp:
(WebCore::DOMCSSNamespace::supports):
* css/DOMCSSNamespace.h:
* css/DOMCSSNamespace.idl:

LayoutTests:

* fast/text/variations/css-supports-runtime-switch-expected.txt: Added.
* fast/text/variations/css-supports-runtime-switch.html: Added.


Identifier: 182080@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@208321 268f45cc-cd09-0410-ab3c-d52691b4dbfc