AX: [ATK] Wrong selected element at a given index in a list box (redux) https://bugs.webkit.org/show_bug.cgi?id=164430
Reviewed by Darin Adler.

Source/WebCore:

This essentially undoes the implementation change resulting from r164577.
As stated in the ATK documentation, atk_selection_ref_selection() takes
"a gint specifying the index in the selection set. (e.g. the ith selection
as opposed to the ith child)." r164577 deliberately modified that, causing
atk_selection_ref_selection() to treat the index as if it were the position
with respect to all of the children. There is different API in ATK, namely
atk_object_ref_accessible_child(), when the ith child from the set of all
children is sought.

Tests: accessibility/aria-listbox-no-selection.html
       accessibility/native-listbox-no-selection.html

* accessibility/atk/WebKitAccessibleInterfaceSelection.cpp:
(optionFromSelection):

LayoutTests:

Add tests to ensure listboxes with no selected children do not report
a selected child. Modify select-element-at-index.html to reflect the
corrected behavior for ATK. Move the Mac port's expectations to the
shared expectations.

* accessibility/aria-listbox-no-selection-expected.txt: Added.
* accessibility/aria-listbox-no-selection.html: Added.
* accessibility/native-listbox-no-selection-expected.txt: Added.
* accessibility/native-listbox-no-selection.html: Added.
* accessibility/select-element-at-index-expected.txt: Modified.
* accessibility/select-element-at-index.html: Modified.
* platform/mac/accessibility/select-element-at-index-expected.txt: Removed.


Identifier: 182174@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@208442 268f45cc-cd09-0410-ab3c-d52691b4dbfc